+++ primitives/src/transaction.rs
@@ -715,6 +715,21 @@ pub enum Legacy {
     Legacy,
 }
 
+impl RichcmpEqualityOnly for Legacy {}
+
+#[pymethods]
+impl Legacy {
+    fn __richcmp__(
+        &self,
+        other: &Self,
+        op: pyo3::basic::CompareOp,
+    ) -> pyo3::prelude::PyResult<bool> {
+        // we override the default impl since it implicitly casts to in which causes problems when transaction
+        // version is represented as `Legacy | int`.
+        solders_traits::RichcmpEqualityOnly::richcmp(self, other, op)
+    }
+}
+
 impl From<Legacy> for LegacyOriginal {
     fn from(x: Legacy) -> Self {
         match x {