+++ src/message.rs
@@ -1,5 +1,8 @@
 use pyo3::{prelude::*, types::PyTuple, PyTypeInfo};
-use solders_primitives::message::{Message, MessageAddressTableLookup, MessageHeader, MessageV0};
+use solders_primitives::message::{
+    from_bytes_versioned, to_bytes_versioned, Message, MessageAddressTableLookup, MessageHeader,
+    MessageV0,
+};
 
 pub(crate) fn create_message_mod(py: Python<'_>) -> PyResult<&PyModule> {
     let m = PyModule::new(py, "message")?;
@@ -7,6 +10,8 @@ pub(crate) fn create_message_mod(py: Python<'_>) -> PyResult<&PyModule> {
     m.add_class::<MessageHeader>()?;
     m.add_class::<MessageV0>()?;
     m.add_class::<MessageAddressTableLookup>()?;
+    m.add_function(wrap_pyfunction!(to_bytes_versioned, m)?)?;
+    m.add_function(wrap_pyfunction!(from_bytes_versioned, m)?)?;
     let typing = py.import("typing")?;
     let union = typing.getattr("Union")?;
     m.add(